requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
capstone==6.0.0a1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
colored-traceback==0.4.2
configparser==4.0.2
cryptography==43.0.3
Django==5.1.7
djangorestframework==3.15.2
fastapi==0.115.12
h11==0.14.0
hexdump==3.3
idna==3.10
intervaltree==3.1.0
Mako==1.3.6
Markdown==3.7
MarkupSafe==3.0.2
more-itertools==5.0.0
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
plumbum==1.9.0
psutil==6.1.0
pwn==1.0
pwntools==4.13.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.0
pydantic_core==2.33.0
pydub==0.25.1
pyelftools==0.31
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-evtx==0.7.4
python-multipart==0.0.20
pytz==2025.2
pywin32==308
requests==2.32.3
ROPGadget==7.5
rpyc==6.0.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.1
trio==0.26.2
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
unicorn==2.1.1
unix-ar==0.2.1
urllib3==2.2.3
uvicorn==0.34.0
websocket-client==1.8.0
wsproto==1.2.0
zipp==1.0.0
zstandard==0.23.0
